+++ contract/contracts/eSpacePoolBridge.sol
@@ -2,9 +2,10 @@
 pragma solidity ^0.8.0;
 
 import "@openzeppelin/contracts/access/Ownable.sol";
-import "../interfaces/ICrossSpaceCall.sol";
-import "../interfaces/IPoSPool.sol";
+import "./interfaces/ICrossSpaceCall.sol";
+import "./interfaces/IPoSPool.sol";
 
+// eSpace pool bridge contract
 contract CoreBridge is Ownable {
   CrossSpaceCall internal crossSpaceCall;