+++ contract/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v1.3.0
 
-This update introduces the function of locking CFX to obtain voting rights for PoS Pools. After obtaining voting rights, you can participate in Conflux on-chain parameters voting and community governance voting. The user can also obtain PoS staking rewards while locking CFX.
+This update introduces the function of locking CFX to obtain voting rights for PoS Pools. After obtaining voting rights, you can participate in Conflux on-chain parameters voting and community governance voting. The means user can obtain PoS staking rewards while participating governance voting.
 
 Check this doc for more details: [v1.3](./docs/v1.3.md)